feat: return to auto schema

+++ src/features/graphql/users/users.resolver.ts
@@ -1,19 +1,29 @@
-import { GraphQLAuth } from '@features/auth/auth.decorators';
+import {
+  GraphQLAuth,
+  GraphQLCurrentUser,
+} from '@features/graphql/auth/auth.decorators';
+import {
+  NotAuthorizedError,
+  ResourcesFilters,
+} from '@features/graphql/common/types';
+import { User } from '@features/graphql/users/entities';
 import {
   CreateUserInput,
   CreateUserPayload,
   DeleteUserInput,
-  GetUserArgs,
-  GetUsersArgs,
+  DeleteUserPayload,
+  GetUserFiltersInput,
+  GetUsersFiltersInput,
   UpdateUserInput,
+  UpdateUserPayload,
   UserEmailAlreadyTakenError,
   UserNotFoundError,
   UserPayload,
   UserResult,
+  UserRole,
   UsersPayload,
 } from '@features/graphql/users/types';
 import { UsersService } from '@features/graphql/users/users.service';
-import { UserRole } from '@generated/graphql.schema';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { generateError, generateResult, GraphQLTNError } from '@utils';
 
@@ -21,27 +31,32 @@ import { generateError, generateResult, GraphQLTNError } from '@utils';
 export class UsersResolver {
   constructor(private readonly usersService: UsersService) {}
 
-  @Query(() => UserPayload, { name: 'user', nullable: true })
+  @Query(() => UserPayload, { name: 'user' })
   @GraphQLAuth(UserRole.USER)
   async getUser(
-    @Args() getUserArgs: GetUserArgs,
+    @Args('getUserFiltersData') getUserFiltersData: GetUserFiltersInput,
   ): Promise<typeof UserPayload | GraphQLTNError> {
-    const user = await this.usersService.getUser(getUserArgs);
+    const user = await this.usersService.getUser(getUserFiltersData);
     if (!user) return generateError(UserNotFoundError.name);
     return generateResult({ user });
   }
 
-  @Query(() => [UsersPayload], { name: 'users', nullable: 'items' })
+  @Query(() => UsersPayload, { name: 'users' })
   @GraphQLAuth(UserRole.ADMIN)
   async getUsers(
-    @Args() getUsersArgs: GetUsersArgs,
+    @Args('getUsersFiltersData') getUsersFiltersData: GetUsersFiltersInput,
+    @Args() { take, skip }: ResourcesFilters,
   ): Promise<typeof UsersPayload | GraphQLTNError> {
-    const users = await this.usersService.getUsers(getUsersArgs);
+    const users = await this.usersService.getUsers({
+      getUsersFiltersData,
+      take,
+      skip,
+    });
 
     return generateResult({ users });
   }
 
-  @Mutation(() => UserResult)
+  @Mutation(() => CreateUserPayload)
   async createUser(
     @Args('createUserData') createUserData: CreateUserInput,
   ): Promise<typeof CreateUserPayload | GraphQLTNError> {
@@ -52,19 +67,33 @@ export class UsersResolver {
     return generateResult({ user });
   }
 
-  @Mutation(() => UserResult)
+  @Mutation(() => UpdateUserPayload)
   @GraphQLAuth(UserRole.USER)
   async updateUser(
     @Args('updateUserData') updateUserData: UpdateUserInput,
-  ): Promise<void | GraphQLTNError> {
-    return this.usersService.updateUser(updateUserData);
+  ): Promise<typeof UpdateUserPayload | GraphQLTNError> {
+    await this.usersService.updateUser(updateUserData);
+
+    const user = this.getUser({ userID: updateUserData.userID });
+
+    return generateResult({ user });
   }
 
-  @Mutation(() => UserResult)
+  @Mutation(() => DeleteUserPayload)
   @GraphQLAuth(UserRole.USER)
   async deleteUser(
     @Args('deleteUserData') deleteUserData: DeleteUserInput,
-  ): Promise<void | GraphQLTNError> {
-    return this.usersService.deleteUser(deleteUserData);
+    @GraphQLCurrentUser() user: User,
+  ): Promise<typeof DeleteUserPayload | GraphQLTNError> {
+    if (
+      !user.roles.includes(UserRole.ADMIN) &&
+      !user.roles.includes(UserRole.MODERATOR) &&
+      user.id !== deleteUserData.userID
+    )
+      return generateError(NotAuthorizedError.name);
+
+    await this.usersService.deleteUser(deleteUserData);
+
+    return generateResult({ isDeleted: true });
   }
 }
